+++ src/platformer/connection/packets/PlayerRespawnPacket.java
@@ -18,7 +18,7 @@ public class PlayerRespawnPacket extends Packet {
     }
 
     @Override
-    protected int getId() {
+    protected byte getId() {
         return 5;
     }
